Retrieve Cryptocurrency Data

Calls the Bitfinex API to retrieve cryptocurrency data for a single ticker.

ticker(SYMBOL = "tBTCUSD")

Arguments

SYMBOL
Character ticker symbol for a single cryptocurrency. Must be prepended with a t or f to specify trading or funding cryptocurrencies, respectively.

Value

A tibble. See details for column names.

Details

tibble descriptions:

SYMBOL: The symbol/ticker of the requested data

FRR: Flash Return Rate - average of all fixed rate funding over the last hour (funding tickers only)

BID: Price of last highest bid

BID_PERIOD: Bid period covered in days (funding tickers only)

BID_SIZE: Sum of the 25 highest bid sizes

ASK: Price of last lowest ask

ASK_PERIOD: Ask period covered in days (funding tickers only)

ASK_SIZE: Sum of the 25 lowest ask sizes

DAILY_CHANGE: Amount that the last price has changed since yesterday

DAILY_CHANGE_RELATIVE: Relative price change since yesterday (*100 for percentage change)

LAST_PRICE: Price of the last trade

VOLUME: Daily volume

HIGH: Daily high

LOW: Daily low

FRR_AMOUNT_AVAILABLE: The amount of funding that is available at the Flash Return Rate (funding tickers only)

Examples

if (FALSE) {
    # Get Bitcoin info
    ticker(SYMBOL = "tBTCUSD")

    # Get USD info
    tickers(SYMBOL = "USD")
}